libAquae: read, parse and validate the metadata
The process of internalising the Federation Metadata is fairly robust but
everything else is quite early. In particular, the exception style needs to
be reworked to use a Aquae-specific exception.

We don't yet read in any of the (Data Sharing) Agreement data from the
Metadata.

testclient (invoked from libAquae/ with './run testclient') will read the
examle Metadata file and tersely summarise what it finds.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>